cms/serializers.py
from rest_framework import serializers
from .models import FAQ, Content


class FAQSerializer(serializers.ModelSerializer):
    class Meta:
        model = FAQ
        fields = ["id", "name", "description"]


class ContentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Content
        fields = ["id", "name", "code", "description"]
